Product handling (#11)
* Modified product handling

* JS at least works

* Upgrades for employee shift handling

* Added КОСТЫЛИ; CSS is now blessed

+++ Web/Requests/ProductRequests/UpdateProductComponentApiRequest.cs
@@ -6,5 +6,5 @@ public class UpdateProductComponentApiRequest
     public required int Product { get; set; }
     public required string Name { get; set; }
     public int? Quantity { get; set; }
-    public float? Weight { get; set; }
+    public double? Weight { get; set; }
 }
